fix: cast picker colors across ModernUI variants

+++ common/src/main/java/indi/etern/musichud/client/ui/components/UniPlaylistPickerModal.java
@@ -332,8 +332,8 @@ public final class UniPlaylistPickerModal {
             int targetName = value ? 0xFF000000 : Theme.NORMAL_TEXT_COLOR;
             int targetCount = value ? 0xFF333333 : Theme.SECONDARY_TEXT_COLOR;
             int initialBackground = currentBackground;
-            int initialName = name.getCurrentTextColor();
-            int initialCount = count.getCurrentTextColor();
+            int initialName = (int) name.getCurrentTextColor();
+            int initialCount = (int) count.getCurrentTextColor();
             currentBackground = targetBackground;
             animation = ValueAnimator.ofFloat(0f, 1f);
             animation.addUpdateListener(valueAnimator -> {
